Refactored some boilerplate code for budgets
