+++ app/src/main/java/com/example/expensetracker/ui/common/ExpenseFAB.kt
@@ -7,7 +7,7 @@ import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import com.example.expensetracker.ui.screen.accounts.AccountsDestination
 import com.example.expensetracker.ui.screen.entities.EntitiesDestination
-import com.example.expensetracker.ui.screen.report.BudgetsDestination
+import com.example.expensetracker.ui.screen.budget.BudgetsDestination
 import com.example.expensetracker.ui.screen.report.ReportsDestination
 import com.example.expensetracker.ui.screen.transactions.TransactionsDestination
