+++ app/src/main/java/com/example/expensetracker/ui/common/ExpenseNavBar.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.unit.dp
 import com.example.expensetracker.ui.navigation.NavigationDestination
 import com.example.expensetracker.ui.screen.accounts.AccountsDestination
 import com.example.expensetracker.ui.screen.entities.EntitiesDestination
-import com.example.expensetracker.ui.screen.report.BudgetsDestination
+import com.example.expensetracker.ui.screen.budget.BudgetsDestination
 import com.example.expensetracker.ui.screen.report.ReportsDestination
 import com.example.expensetracker.ui.screen.settings.SettingsDestination
 import com.example.expensetracker.ui.screen.transactions.TransactionsDestination
